requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.6
annotated-types==0.7.0
attrs==25.3.0
cffi==1.17.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
idna==3.10
jax==0.6.0
jaxlib==0.6.0
Jinja2==3.1.6
kiwisolver==1.4.8
lightning-utilities==0.14.3
MarkupSafe==3.0.2
matplotlib==3.10.3
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.2.1
propcache==0.3.1
protobuf==4.25.7
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyqtgraph==0.13.7
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1.post0
PyYAML==6.0.2
scipy==1.15.3
screeninfo==0.8.1
sentencepiece==0.2.0
setuptools==80.3.1
simsimd==6.2.1
six==1.17.0
sounddevice==0.5.1
stringzilla==3.12.5
sympy==1.14.0
torch==2.7.0
torchmetrics==1.7.1
torchvision==0.22.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
yarl==1.20.0
